% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi_generators.R
\name{generate_foi}
\alias{generate_foi}
\title{\code{generate_foi} get the Force of Infection (FOI) of the the Helicobacter pylori
of the United States of America. This FOI is based in the surveys reported in
the National Health and Nutrition Examination (NHANES) of 1988-1994
and 1999-2000}
\usage{
generate_foi(
  age = c(0:100),
  sex = c("Male", "Female"),
  race = c("Non-Hispanic White", "Non-Hispanic Black", "Other Hispanic",
    "Mexican-American", "Other"),
  period = 1991:2010
)
}
\arguments{
\item{age}{Numeric vector specifying the age of the person with
Helicobacter pylori}

\item{sex}{Character vector specifying the sex of the person (Male, Female)}

\item{race}{Character vector specifying the race of the person
(Non-Hispanic White, Non-Hispanic Black, Other Hispanic, Mexican-American, Other)}

\item{period}{Character vector specifying the period of interest of the FOI. The default period is the year 1991}
}
\value{
A dataframe with the distributions of the alpha and gamma parameters according to the specified variables.
}
\description{
\code{generate_foi} get the Force of Infection (FOI) of the the Helicobacter pylori
of the United States of America. This FOI is based in the surveys reported in
the National Health and Nutrition Examination (NHANES) of 1988-1994
and 1999-2000
}
\examples{
generate_foi(age = c(1,5,80), sex = c("Male", "Female"), race = "Non-Hispanic White", 
period = "2000")

generate_foi(age = 90, sex = "Male", race = "Other", 
period = "1995")

}
