% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anxiliary_functions.R
\name{ilogit}
\alias{ilogit}
\title{Inverse logit unction}
\usage{
ilogit(logodds)
}
\arguments{
\item{logodds}{log-odds}
}
\value{
The probability of log-odds.
}
\description{
\code{ilogit} computes the probability \code{p} from a log-odds.
}
