% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anxiliary_functions.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{probability}
}
\value{
The log-odds of \code{p}.
}
\description{
\code{logit} computes the log odds of a probability \code{p}.
}
