% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi_plotting.R
\name{plot_foi_shiny}
\alias{plot_foi_shiny}
\title{\code{plot_foi_shiny} Plot the Force of Infection (FOI)
according to the variables specified by the user.}
\usage{
plot_foi_shiny(
  age = c(0:100),
  sex = c(),
  race = c("Non-Hispanic White", "Non-Hispanic Black", "Other Hispanic",
    "Mexican-American", "Other"),
  period = 1991:2010,
  percs = T,
  age_bracket = 10,
  save_data = F,
  save_plot = F,
  plot_name = "FOI",
  data_name = "df_foi",
  width = 10,
  height = 8
)
}
\arguments{
\item{age}{Numeric vector specifying the age of the person with
Helicobacter pylori}

\item{sex}{Character vector specifying the sex of the person (Male, Female)}

\item{race}{Character vector specifying the race of the person
(Non-Hispanic White, Non-Hispanic Black, Other Hispanic, Mexican-American, Other)}

\item{period}{Character vector specifying the period of interest of the FOI. The default period is the year 1991}

\item{percs}{Logical statement specifying the Confidence Intervals (Upper Bound (UB) and Lower Bound (LB)).
TRUE to get Confidence Intervals.}

\item{age_bracket}{Set the number of brakes for the age in the x axis in the plot.}

\item{save_data}{Logical statement. TRUE to save the dataframe of the plot.}

\item{save_plot}{Logical statement. TRUE to save the plot as a jpg file.}

\item{plot_name}{Name in string that will have the saved plot.}

\item{data_name}{Name in string that will have the saved dataframe.}

\item{width}{width of the saved plot.}

\item{height}{height of the saved plot.}
}
\value{
A dataframe with the Force of Infection (FOI) according to the specified variables by the user.
}
\description{
\code{plot_foi_shiny} Plot the Force of Infection (FOI)
according to the variables specified by the user.
}
\examples{
plot_foi(age = c(1,5,80), sex = c("Male", "Female"), race = "Non-Hispanic White", 
period = "2000", percs = F, age_bracket = 5, save_data = F, save_plot = T,
plot_name = "Force of Infection", data_name = "df_force_infection", 
width = 14, height = 10)

plot_foi(age = 90, sex = "Male", race = c("Non-Hispanic White", "Non-Hispanic Black", 
"Other Hispanic", "Mexican-American", "Other"), 
period = c(1995, 2000, 2010))

}
