% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi_generators.R
\name{summary_foi}
\alias{summary_foi}
\title{\code{summary_foi} summarise the Force of Infection (FOI) in a dataframe
according to the variables specified by the user.}
\usage{
summary_foi(
  age = c(0:100),
  sex = c("Male", "Female"),
  race = c("Non-Hispanic White", "Non-Hispanic Black", "Other Hispanic",
    "Mexican-American", "Other"),
  period = 1991:2010,
  percs = T
)
}
\arguments{
\item{age}{Numeric vector specifying the age of the person with
Helicobacter pylori}

\item{sex}{Character vector specifying the sex of the person (Male, Female)}

\item{race}{Character vector specifying the race of the person
(Non-Hispanic White, Non-Hispanic Black, Other Hispanic, Mexican-American, Other)}

\item{period}{Character vector specifying the period of interest of the FOI. The default period is the year 1991}

\item{percs}{Logical statement specifying the Confidence Intervals (Upper Bound (UB) and Lower Bound (LB)).
TRUE to get Confidence Intervals.}
}
\value{
A dataframe with the Force of Infection (FOI) according to the specified variables by the user.
}
\description{
\code{summary_foi} summarise the Force of Infection (FOI) in a dataframe
according to the variables specified by the user.
}
\examples{
summary_foi(age = c(1,5,80), sex = c("Male", "Female"), race = "Non-Hispanic White", 
period = "2000", percs = c(LB, UB))

summary_foi(age = 90, sex = "Male", race = "Other", 
period = "1995", percs = c(LB, UB))

}
